reconfig diesel

+++ pm-core/src/db/schema/theme.rs
@@ -2,7 +2,7 @@
 
 diesel::table! {
     color_scheme (id) {
-        id -> BigInt,
+        id -> Integer,
         name -> Text,
         primary -> Text,
         secondary -> Text,
@@ -16,11 +16,14 @@ diesel::table! {
 
 diesel::table! {
     theme (id) {
-        id -> BigInt,
-        color_scheme_id -> BigInt,
+        id -> Integer,
+        color_scheme_id -> Integer,
     }
 }
 
 diesel::joinable!(theme -> color_scheme (color_scheme_id));
 
-diesel::allow_tables_to_appear_in_same_query!(color_scheme, theme,);
+diesel::allow_tables_to_appear_in_same_query!(
+    color_scheme,
+    theme,
+);
